add fixes for xl colab updates

+++ src/mpu/transformer.py
@@ -125,9 +125,9 @@ class GPT3ParallelSelfAttention(torch.nn.Module):
 
         if self.use_deepspeed_sparse:
             context_layer = self.sparse_self_attention(
-                query_layer.float(),
-                key_layer.float(),
-                value_layer.float(),
+                query_layer.half(),
+                key_layer.half(),
+                value_layer.half(),
                 attn_mask=ltor_mask).half()
         else:
             # Raw attention scores. [b, np, s, s]
